New versions of clay, oometa, dicttool, and tool.
Refactored all three packages to allow existing tool libraries to
leverage the clay framework without having to participate in the clay 
framework. Fixed a lot of hairy issues dealing with the opposite
ways in which oo::meta and dicttool::merge/clay approach merging dicts.

All tests now pass. Some esoteric tests in tool have been removed. The 
answers they were producing in the past were a shade of wrong, we were
just documenting the wrongness.

The major change for tool 0.8 is that it is aware of clay, and it
now utilizes the same metadata structure elements as an equilent
clay class/object. We are just preserving the notation.

Tool objects now re-build their ensembles on mixin using an
adapted vesion of the facilities from clay.

An object should be able to inherit ancestors from either and still work perfectly.

(Though I really should put some tests together to demonstrate that...)

Eliminiated dicttool::canonical. We really only need to know where a value
is going to be stored in the end.