Added an annotation capacity to practcl, which allows the doctools generator to read and document class variables, options, and delegates.
Added missing documentation to httpd.

New version of clay which adds a new "branch" method to oo::class/oo::object's
clay ensemble. The branch method tells the system to mark the designated
address as a branch, even it empty.

Fixed a bug in clay where a Dict or Array keyword with no values
would fail to actually register in the clay system.